factored out cursor struct and co
